Trying something different. Adding a seperate file for my user task.
